feat: correcciones de PR
